# PERF: Improve perf initalizing DataFrame with a range