@topper-123 2019-12-09
> 
> 
> Can you check we already have a test for this construction?

It's tested in ``pandas\tests\frame\test_constructors.py`` line 1079 and used in a few more places.