@topper-123 2019-12-09
> nice. does this need/merit an asv?

IMO that would not be really important, though I'm not against it either. This pattern is probably mostly used interactively, in ipython and jupyter etc. and little otherwise.