@topper-123 2019-12-10
I've added a ASV. I can take on #30173 in a followup for both Series and Dataframes.